t24_complete_system/backend/requirements.txt
fastapi==0.111.0
uvicorn==0.27.1
sqlalchemy==2.0.27
pydantic==1.10.12
psycopg[binary]==3.2.2
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.2
python-multipart==0.0.9
email-validator==2.1.0
python-dotenv==1.0.1
alembic==1.13.1
PyJWT==2.8.0
psycopg2-binary>=2.9.7
pandas>=2.0.3
requests>=2.31.0
